Extract to getDatabases()
