+++ src/index.ts
@@ -79,35 +79,14 @@ app.action("open-modal-button", async({ ack, body, client, logger}) => {
   // console.dir(body, {depth: null})
 
   try {
-    const dbs = await notion.searchDb();
-    const dbChoices = []
-    for (const db of dbs) {
-      if (db.object != "database") {
-        continue
-      }
-      if (!isFullDatabase(db)) {
-        continue
-      }
-      if (db.title[0] == null) {
-        continue
-      }
-      dbChoices.push({
-        title: db.title[0].plain_text,
-        value: db.id,
-      })
-    }
-    const sortedDbChoices = dbChoices.sort((a,b)=> {
-      return a.title.localeCompare(b.title)
-    })
-    console.log(sortedDbChoices)
-
+    const dbs = await notion.getDatabases()
     const metaData = {
       channel_id: body.channel.id,
       thread_ts: body.message.thread_ts,
     }
     await client.views.open({
       trigger_id: body.trigger_id,
-      view: slack.searchDbView(metaData, sortedDbChoices),
+      view: slack.searchDbView(metaData, dbs),
     })
   } catch (error) {
     logger.error(error)
@@ -164,29 +143,11 @@ app.action('change_db-action', async({ack, body, client, logger}) => {
     const pm = JSON.parse(body.view.private_metadata)
     console.dir(pm, {depth: null})
 
-    const dbs = await notion.searchDb();
-    const dbChoices = []
-    for (const db of dbs) {
-      if (db.object != "database") {
-        continue
-      }
-      if (!isFullDatabase(db)) {
-        continue
-      }
-      dbChoices.push({
-        title: db.title.length > 0 ? db.title[0].plain_text : "Untitled",
-        value: db.id,
-      })
-    }
-    const sortedDbChoices = dbChoices.sort((a,b)=> {
-      return a.title.localeCompare(b.title)
-    })
-    console.log(sortedDbChoices)
-
+    const dbs = await notion.getDatabases()
     await client.views.update({
       view_id: body.view.id,
       hash: body.view.hash,
-      view: slack.searchDbView(pm, sortedDbChoices),
+      view: slack.searchDbView(pm, dbs),
     })
   } catch (error) {
     logger.error(error)
