+++ src/notion.ts
@@ -1,4 +1,7 @@
-import { Client } from '@notionhq/client'
+import {
+    Client,
+    isFullDatabase,
+  } from '@notionhq/client'
 import {
   QueryDatabaseParameters,
   QueryDatabaseResponse,
@@ -380,4 +383,25 @@ export const getSelectedDbPropValues = async (
     }
   })
   return props
+}
+
+export const getDatabases = async () => {
+  const dbs = await searchDb();
+  const dbChoices = []
+  for (const db of dbs) {
+    if (db.object != "database") {
+      continue
+    }
+    if (!isFullDatabase(db)) {
+      continue
+    }
+    dbChoices.push({
+      title: db.title.length > 0 ? db.title[0].plain_text : "Untitled",
+      value: db.id,
+    })
+  }
+  const sortedDbChoices = dbChoices.sort((a,b)=> {
+    return a.title.localeCompare(b.title)
+  })
+  return sortedDbChoices
 }
